requirements.txt
git+https://github.com/tensorpack/dataflow.git@4ac75d6b000c887b68bbc4ace11c57a47eff662c#egg=dataflow
docutils==0.16
flake8==3.9.0
lmdb==1.1.1
matplotlib==3.4.1
mypy==0.812
numpy==1.20.2
open3d==0.12.0
opencv-python==4.5.1.48
pandas==1.2.4
pybind11==2.6.2
pykitti==0.3.1
pytest==6.2.3
python-prctl==1.8.1
pytorch-ignite==0.4.4
PyYAML==5.4.1
git+https://github.com/lessw2020/Ranger-Deep-Learning-Optimizer@a170e3dac6443763961d0cb7a16294e4f45f93e4#egg=ranger
recommonmark==0.7.1
scipy==1.6.2
Sphinx==3.5.4
sphinx-rtd-theme==0.5.2
tensorboardX==2.2
torch==1.8.1
torch-cluster==1.5.9
torchgeometry==0.1.2
torchvision==0.9.1
transforms3d==0.3.1
vtk==9.0.1
protobuf==3.20.*

traitlets==5.5.0
Jinja2==3.0
ipython==7.23.1

# ipykernel 6.17.1 requires ipython>=7.23.1, but you'll have ipython 7.21.0 which is incompatible.
# ipykernel 6.17.1 requires traitlets>=5.1.0, but you'll have traitlets 5.0.5 which is incompatible.
# nbformat 5.7.0 requires traitlets>=5.1, but you'll have traitlets 5.0.5 which is incompatible.
# nbclient 0.7.0 requires traitlets>=5.2.2, but you'll have traitlets 5.0.5 which is incompatible.
# nbconvert 7.2.5 requires jinja2>=3.0, but you'll have jinja2 2.11.3 which is incompatible.
# nbconvert 7.2.5 requires markupsafe>=2.0, but you'll have markupsafe 1.1.1 which is incompatible.
# jupyter-server 1.23.3 requires traitlets>=5.1, but you'll have traitlets 5.0.5 which is incompatible.
